# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Farmart Grocery Store API",
    "description": "API Specification ออกแบบจากฟังก์ชันและหน้าเว็บทั้งหมดของระบบ Farmart Supermarket / E-Commerce (Next.js App Router)\n\nครอบคลุมทั้งระบบแค็ตตาล็อกสินค้า (Products), หมวดหมู่และแบรนด์ (Categories & Brands), ดีลพิเศษประจำวัน (Top Saver, Best Seller, Just Landing), ระบบรีวิวและการให้คะแนน (Reviews), ตะกร้าสินค้า (Cart), รายการสินค้าที่ชอบ (Wishlist), จดหมายข่าว (Newsletter), และสถานะระบบ (Health Check)",
    "version": "1.0.0",
    "contact": {
      "name": "Farmart Engineering Team",
      "email": "api-support@farmart.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "/",
      "description": "Current Origin Server"
    }
  ],
...
  "paths": {
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "ค้นหารายการสินค้า (Search & Filter Products)",
        "description": "ดึงรายการสินค้าพร้อมรองรับการค้นหาตามคีย์เวิร์ด, กรองตามหมวดหมู่, แบรนด์, สินค้าออร์แกนิก และการจัดเรียง",
        "operationId": "getProducts",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "คำค้นหาชื่อสินค้า แบรนด์ หรือรายละเอียด",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "avocado"
          },
          {
            "name": "category",
            "in": "query",
            "description": "Slug หรือชื่อหมวดหมู่สินค้า",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "fruits-vegetables"
          },
          {
            "name": "brand",
            "in": "query",
            "description": "ชื่อแบรนด์สินค้า",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Farmart Organic Direct"
          },
          {
            "name": "isOrganic",
            "in": "query",
            "description": "กรองเฉพาะสินค้าที่เป็นออร์แกนิก 100%",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "example": true
          },
          {
            "name": "sort",
            "in": "query",
            "description": "การเรียงลำดับสินค้า",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["featured", "price_asc", "price_desc", "rating", "newest"],
              "default": "featured"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "หมายเลขหน้าที่ต้องการดึงข้อมูล",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "จำนวนสินค้าต่อหน้า",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 12,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ดึงรายการสินค้าสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          },
          "400": {
            "description": "พารามิเตอร์การค้นหาไม่ถูกต้อง",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProductCardItem": {
        "type": "object",
        "required": ["id", "name", "price", "rating", "reviewCount"],
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Fresh Organic Hass Avocado (Pack of 4)"
          },
          "brand": {
            "type": "string",
            "example": "Farmart Organic Direct"
          },
          "category": {
            "type": "string",
            "example": "Fruits & Vegetables"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 6.49
          },
          "oldPrice": {
            "type": "number",
            "format": "float",
            "example": 8.99
          },
          "unit": {
            "type": "string",
            "example": "4 pcs (~650g)"
          },
          "discountPercent": {
            "type": "integer",
            "example": 28
          },
          "isOrganic": {
            "type": "boolean",
            "example": true
          },
          "rating": {
            "type": "number",
            "format": "float",
            "example": 4.9
          },
          "reviewCount": {
            "type": "integer",
            "example": 128
          },
          "imageUrl": {
            "type": "string",
            "example": "/images/products/avocado.png"
          }
        }
      },
...
      "ProductListResponse": {
        "type": "object",
        "required": ["total", "page", "limit", "products"],
        "properties": {
          "total": {
            "type": "integer",
            "example": 48
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 12
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductCardItem"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": {
            "type": "string",
            "example": "Not Found"
          },
          "message": {
            "type": "string",
            "example": "Product with ID 999 not found"
          },
          "statusCode": {
            "type": "integer",
            "example": 404
          }
        }
      }
    }
  }
}